requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.5.1
Django==2.1.5
django-debug-toolbar==1.11
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-bulk==0.2.1
docopt==0.6.2
flake8==3.6.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
model-mommy==1.6.0
more-itertools==5.0.0
mysqlclient==1.3.14
pluggy==0.8.1
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.7
requests==2.20.1
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.0
